move regulation-related commands from base to regulations
